BBM - version 2.0 released
> Disclaimer:
- The Bb Codes manager didn't change at all
- The button manager dropped the support of TinyMCE 3.x
- The button manager is only compatible with TinyMCE 4.x (see the addon
TinyMCE Quattro)
- After having installed this version, you should be able to revert to
version 1.x (not supported)
- The BbCodes xml files have changed a little to add new options
- If a major problem is found on the version 1 of the BBM, it will be
updated on github. If users don't feel confortable to download files
there, I will create an archive and upload it inside a post.
- I didn't have time yet to recreate the UI of the Advanced Bb Codes.
That will be my next move.

> New methods to create a mce 4 button:
- It can be a blank button (you can use this way png icons - I don't
support this, see the above options to understand why)
- It can be a text button, just  type a text, you've got a text button.
Time to create this button: 3 seconds (if you type slow).
- It can be a font icon button based on the TinyMCE official icons set
or the XenForo icons set. I will explain this in a tutorial. Type to
create this kind of button: 30 seconds (if you are slow to select which
kind of button you want to use)

>New return methods for the Button:
- The "Direct" method will insert the Bb Code with options and content
if needed
- The "Template" method will load an overlay window  from a custom
template of your choice. This template must have a Javascript to manage
the submit input. The framework of TinyMCE Quattro includes has been
developped to make your life easy: even if you are a Javascript beginner
you will be able to use this. An example is provided; see the template
"quattro_dialog_bbm_demo". It is full documented.

> The Buttons manager now deals with icons. It makes the configuration
easier

> The Buttons manager has now a reset option to help you to get back the
default layout.

#Update from 1.x version#
> Like usual, nothing has changed
> Use the reset function to recreate your editor layout
> If you want, you can delete all orphan buttons (from the Enhanced
TinyMCE addon) - they have a purple color.

Have fun with this!