Version 2.3.0 Preview
It's now possible to override the editor type (XenForo Redactor or
TinyMCE Quattro) by Controller/View or forums. In simple words, if you
want to use a custom config of MCE4 in a forum (or by Controller/View)
whereas your default configuration is to use the XenForo default editor,
you can now. The opposite is also possible.