Version 3.1.0 released
- Xon added a nice cache system for the tags map. It must be enabled
inside the options.
Explanation: when a page is parsed, the Bbm creates a map of all the
tags with their personal data. This operation can be intensive (it
depends on how many Bb Codes there are), the new system allows to cache
them.
- For Bb Code developers (from a Xon idea): add a uniqid function in the
formatter (can be called from the callback).
Purpose: avoid to use the intensively the php uniqid function. Source:
http://www.ning.com/code/2012/02/uniqid-is-slow
- Hide XenForo button when its XenForo Bb Code parsing permission (based
on the Visitor object) is disabled. Only works with the following Bb
Codes : email, url, img & media. The "targeted nodes" option is ignored
(can get node information in this part of the script).
Xon is now a contributor on the Github branch, which means he should now
have access to version on Github and update it if needed. You can fully
trust his commits: he's serious & reactive, knows how to code very well
and always take care of improving performance. A big thank to him.